# docs/api_specs.yaml
openapi: 3.0.3
info:
    title: Echo API
    version: 1.0.0
    description: |
        Local-first REST API for Echo - a multimodal semantic second brain.
        All processing runs on the user's device. No external network calls are
        made during normal operation. The API never generates content; it only
        stores, retrieves and surfaces the user's own existing notes.
servers:
    - url: http://127.0.0.1:8000/
      description: Local development server

#  REUSABLE COMPONENTS
components:
    schemas:
...
        JobType:
            type: string
            enum: [transcribe, caption, ocr, embed, generate_links]
        JobStatus:
            type: string
            enum: [queued, running, complete, failed]
...
        Job:
            type: object
            required: [id, job_type, status, retry_count, created_at]
            description: >
                Exactly one of note_id / attachment_id is set, never both,
                never neither. "embed" and "generate_links" jobs reference
                note_id; "caption", "ocr" and "transcribe" jobs reference
                attachment_id. Enforced at the database layer via CHECK
                constraint.
            properties:
                id:
                    type: string
                    format: uuid
                note_id:
                    type: string
                    format: uuid
                    nullable: true
                attachment_id:
                    type: string
                    format: uuid
                    nullable: true
                job_type:
                    $ref: "#/components/schemas/JobType"
                status:
                    $ref: "#/components/schemas/JobStatus"
                error_message:
                    type: string
                    nullable: true
                retry_count:
                    type: integer
                    minimum: 0
                    default: 0
                created_at:
                    type: string
                    format: date-time
                started_at:
                    type: string
                    format: date-time
                    nullable: true
                completed_at:
                    type: string
                    format: date-time
                    nullable: true
...
        Error400:
            type: object
            required: [error, detail]
            properties:
                error:
                    type: string
                    example: "validation_error"
                detail:
                    type: string
                    example: "note_type must be one of: text, image, audio, canvas"
        Error404:
            type: object
            required: [error, detail]
            properties:
                error:
                    type: string
                    example: "not_found"
                detail:
                    type: string
                    example: "Note with id 'a1b2c3d4...' does not exist."
...
        Error500:
            type: object
            required: [error, detail]
            properties:
                error:
                    type: string
                    example: "internal_error"
                detail:
                    type: string
                    example: "An unexpected error occurred."

    #  PARAMETERS
    parameters:
...
        NotesLimitParam:
            name: limit
            in: query
            required: false
            schema:
                type: integer
                minimum: 1
                maximum: 200
                default: 50
            description: >
                Maximum number of notes to return per page. Default of 50
                is intended for general note browsing, not similarity ranking.
...
paths:
...
    /jobs:
        get:
            operationId: listJobs
            summary: List background jobs
            description: >
                Returns all jobs ordered by created_at descending. Used by the
                process queue panel (!F-52) to show running, pending and failed jobs.
            tags: [Jobs]
            parameters:
                - name: status
                  in: query
                  required: false
                  schema:
                      $ref: "#/components/schemas/JobStatus"
                  description: Filter by job status.
                - $ref: "#/components/parameters/NotesLimitParam"
            responses:
                "200":
                    description: Array of jobs.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Job"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
    /jobs/{id}/retry:
        post:
            operationId: retryJob
            summary: Retry a failed job
            description: >
                Re-queues a job that has status "failed". Increments retry_count.
                Returns 400 if the job is not in a failed state.
            tags: [Jobs]
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                "200":
                    description: Job re-queued successfully.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Job"
                "400":
                    description: Job is not in a failed state.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error400"
                "404":
                    description: Job not found.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error404"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error500"
...